[Designer] fix - Scenario variables are wrongly restored if they were modified without saving after Play
Also fix a warning on mac

Tracker Reference: [CORE-1047](https://jira.mensiatech.com/browse/CORE-1047)

+++ plugins/examples/src/box-algorithms/ovpCBoxAlgorithmModifiableSettings.cpp
@@ -33,7 +33,7 @@ bool CBoxAlgorithmModifiableSettings::processClock(OpenViBE::Kernel::IMessageClo
 	//print settings values
 	for(size_t i=0; i < m_SettingsValue.size(); ++i)
 	{
-		this->getLogManager() << LogLevel_Info << "Setting " << i << " value is " << m_SettingsValue[i] << "\n";
+		this->getLogManager() << LogLevel_Info << "Setting " << static_cast<uint64_t>(i) << " value is " << m_SettingsValue[i] << "\n";
 	}
 	this->getLogManager() << LogLevel_Info << "\n";
